% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/boynton.R
\name{boynton}
\alias{boynton}
\title{Retrieve one or more of Boynton's 11 color specifications as hex,
   rgb triplets or a data.frame}
\usage{
boynton(index, ...)
}
\arguments{
\item{index}{a vector of indices to retrieve if missing then all 11 are returned
Boynton's colors are named, so it is possible to specify by name or position index}

\item{...}{further arguments for \code{\link{get_lut}}}
}
\value{
color values in the form specified
\describe{
   \item{hex}{a character vector of hexidecimal strings, this is the default}
   \item{rgb}{a 3-column matrix of red, green and blue values}
   \item{data.frame}{a 4 column data frame, see \code{\link{BOYNTONLUT}} }
}
}
\description{
Retrieve one or more of Boynton's 11 color specifications as hex,
   rgb triplets or a data.frame
}
\examples{
plot(3:11, pch = 19, col = boynton(3:11))
}

