% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/catacolors-package.R
\docType{data}
\name{BOYNTONLUT}
\alias{BOYNTONLUT}
\title{Boynton's 11 colors specified in hexidecimal strings and rgb.}
\format{A 4 variable 11 row data frame color specifications containing
   \describe{
      \item{hex}{hexidecimal string color value}
      \item{red}{Red value scaled 0-255}
      \item{green}{Green value scaled 0-255}
      \item{blue}{Blue value scaled 0-255}
   }}
\source{
adapted from \href{http://stackoverflow.com/questions/470690/how-to-automatically-generate-n-distinct-colors}{Ohad Schneider}
}
\usage{
BOYNTONLUT
}
\description{
Boynton's 11 colors specified in hexidecimal strings and rgb.
}
\keyword{datasets}

