% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/catacolors-package.R
\docType{package}
\name{catacolors}
\alias{catacolors}
\alias{catacolors-package}
\title{Color tables for displaying catagorical data including Glasbey's (256 colors), Wad's (16 colors), Kelly's (22 colors) and Boynton's (11 colors)}
\description{
Color tables for displaying catagorical data including Glasbey's (256 colors), Wad's (16 colors), Kelly's (22 colors) and Boynton's (11 colors)
}

