% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{get_lut}
\alias{get_lut}
\title{Retrieve one or more of color specifications as hex, rgb triplets or a data.frame}
\usage{
get_lut(LUT, index, form = c("hex", "rgb", "data.frame")[1], name = FALSE)
}
\arguments{
\item{LUT}{a color look-up table data frame.  See \code{link{GLASBEYLUT}},
\code{link{WADLUT}}, \code{link{KELLYLUT}} and \code{link{BOYNTONLUT}}.}

\item{index}{a vector of numeric or character indices to retrieve colors,
if missing then all are returned. If the look-up table has row names, then
one can use those for indexing.}

\item{form}{character, either 'hex', 'rgb' or 'data.frame'}

\item{name}{logical, ignored if form is 'data.frame'}
}
\value{
color values in the form specified
\describe{
   \item{hex}{a character vector of hexidecimal strings, this is the default}
   \item{rgb}{a 3-column matrix of red, green and blue values}
   \item{data.frame}{a 4 column data frame, see \code{\link{GLASBEYLUT}} }
}
}
\description{
Retrieve one or more of color specifications as hex, rgb triplets or a data.frame
}
\examples{
lut <- get_lut(WADLUT, 3:16, form = 'hex', name = TRUE)
plot(3:16, pch = 19, col = lut, cex = 3)
}

