% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glasbey-package.R
\docType{package}
\name{glasbey-package}
\alias{glasbey-package}
\title{Discrete 256-color table devised by Glasbey et al suitable for catagorical coloring.}
\description{
Glasbey, Chris, et al. "Colour displays for categorical images."
Color Research & Application 32.4 (2007): 304-309.
}
\details{
Here is an implementation in ImageJ by Gabrial Landini
\url{http://www.dentistry.bham.ac.uk/landinig/software/software.html}
}

