% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glasbey.R
\name{glasbey}
\alias{glasbey}
\title{Retrieve one or more color specifications as hex, rgb triplets or a data.frame}
\usage{
glasbey(index, form = c("hex", "rgb", "data.frame")[1])
}
\arguments{
\item{index}{a vector of indices to retrieve if missing then all 256 are returned}

\item{form}{character specifies return type as 'hex', 'rgb' or 'data.frame'}
}
\value{
color values in the form specified
\describe{
   \item{hex}{a character vector of hexidecimal strings, this is the default}
   \item{rgb}{a 3-column matrix of red, green and blue values}
   \item{data.frame}{a 4 column data frame, see \code{\link{GLASBEYLUT}} }
}
}
\description{
Retrieve one or more color specifications as hex, rgb triplets or a data.frame
}
\examples{
plot(1:10, pch = 19, col = glasbey(11:20))
}

