% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glasbey-package.R
\docType{data}
\name{GLASBEYLUT}
\alias{GLASBEYLUT}
\title{Glasbey's colors specified in hexidecimal strings and rgb}
\format{A 4 variable 256 row data frame color specifications containing
   \describe{
      \item{hex}{hexidecimal string color value}
      \item{red}{Red value scaled 0-255}
      \item{green}{Green value scaled 0-255}
      \item{blue}{Blue value scaled 0-255}
   }}
\source{
adapted from \href{http://www.dentistry.bham.ac.uk/landinig/software/software.html}{Gabrial Landini's} work with ImageJ
}
\usage{
GLASBEYLUT
}
\description{
Glasbey's colors specified in hexidecimal strings and rgb
}
\keyword{datasets}

