% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kelly.R
\name{kelly}
\alias{kelly}
\title{Retrieve one or more of Kelly's 22 color specifications as hex,
   rgb triplets or a data.frame}
\usage{
kelly(index, ...)
}
\arguments{
\item{index}{a vector of indices to retrieve if missing then all 22 are returned
Watlingtons colors are named, so it is possible to specify by name or position index}

\item{...}{further arguments for \code{\link{get_lut}}}
}
\value{
color values in the form specified
\describe{
   \item{hex}{a character vector of hexidecimal strings, this is the default}
   \item{rgb}{a 3-column matrix of red, green and blue values}
   \item{data.frame}{a 4 column data frame, see \code{\link{KELLYLUT}} }
}
}
\description{
Retrieve one or more of Kelly's 22 color specifications as hex,
   rgb triplets or a data.frame
}
\examples{
plot(2:22, pch = 19, col = kelly(2:22))
}

