% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/glasbey.R
\name{show_glasbey}
\alias{show_glasbey}
\title{Show the Glasbey et al colors in a palette view}
\usage{
show_glasbey(label = TRUE)
}
\arguments{
\item{label}{logical, if TRUE then label each color with it's number}
}
\description{
Show the Glasbey et al colors in a palette view
}

