% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/catecolors-package.R
\docType{data}
\name{WADLUT}
\alias{WADLUT}
\title{Watlington's (Wad's) 16 colors specified in hexidecimal strings and rgb.}
\format{A 4 variable 16 row data frame color specifications containing
   \describe{
      \item{hex}{hexidecimal string color value}
      \item{red}{Red value scaled 0-255}
      \item{green}{Green value scaled 0-255}
      \item{blue}{Blue value scaled 0-255}
   }}
\source{
adapted from \href{http://alumni.media.mit.edu/~wad/color/palette.html}{John A Watlinton}
}
\usage{
WADLUT
}
\description{
Watlington's (Wad's) 16 colors specified in hexidecimal strings and rgb.
}
\keyword{datasets}

