% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/catecolors-package.R
\docType{package}
\name{catecolors}
\alias{catecolors}
\alias{catecolors-package}
\title{Color tables for displaying categorical data including Glasbey's (256 colors), Wad's (16 colors), Kelly's (22 colors) and Boynton's (11 colors)}
\description{
Color tables for displaying categorical data including Glasbey's (256 colors), Wad's (16 colors), Kelly's (22 colors) and Boynton's (11 colors)
}

